Extract chart layout logic to a custom hook
This extracts the overly complex logic calculating the chart padding and axis layout variables to a custom hook (`useChartLayout`) from `ChartContainer`.
This significantly improves readability and maintainability.
